refactor: move common constants into seperate file, hide API config on /ask-ripeseed
